+++ source/plugin/PluginVst2xHostCallback.cpp
@@ -115,7 +115,7 @@ VstIntPtr VSTCALLBACK pluginVst2xHostCallback(AEffect *effect, VstInt32 opcode, 
   // This string is used in a bunch of logging calls below
   PluginVst2xId pluginId;
   if(effect != NULL) {
-    pluginId = newPluginVst2xIdWithId(effect->uniqueID);
+    pluginId = newPluginVst2xIdWithId((unsigned long)effect->uniqueID);
   }
   else {
     // During plugin initialization, the dispatcher can be called without a
